[#546] Added error and log handling for all scripts and functions

+++ auto-setup/core/gopass/post_install_scripts/loadInitialPasswordsAndCleanup.sh
@@ -1,26 +1,29 @@
 #!/bin/bash
-set -euo pipefail
 
-# Get hash passed in from the Jenkins based launcher
-hash="$(/usr/bin/sudo cat /var/tmp/.hash)"
+if [[ -f /var/tmp/.hash ]] && [[ -f ${sharedKxHome}/.config/.vmCredentialsFile ]]; then
 
-# Ensure no Windows characters blocking decryption
-/usr/bin/sudo apt-get install dos2unix
-/usr/bin/sudo dos2unix ${sharedKxHome}/.config/.vmCredentialsFile
+       # Get hash passed in from the Jenkins based launcher
+       hash="$(/usr/bin/sudo cat /var/tmp/.hash)"
 
-# Loop through encrypted credentials file and load into GoPass
-while IFS='' read -r credential || [[ -n "$credential" ]]; do
-    name=$(echo ${credential} | cut -f 1 -d':')
-    secret=$(echo ${credential} | cut -f 2 -d':')
-    if [[ "${secret}" != '""' ]] && [[ "${secret}" != "" ]]; then
-           decryptedSecret=$(echo "${secret}" | openssl enc -aes-256-cbc -pbkdf2 -salt -A -a -pass pass:${hash} -d)
-           cleanedSecret=$(cleanOutput "${decryptedSecret}")
-    else
-           decryptedSecret=""
-    fi
-    pushPassword "${name}" "${cleanedSecret}" "base-technical-credentials"
-done < "${sharedKxHome}/.config/.vmCredentialsFile"
+       # Ensure no Windows characters blocking decryption
+       /usr/bin/sudo apt-get install dos2unix
+       /usr/bin/sudo dos2unix ${sharedKxHome}/.config/.vmCredentialsFile
 
-# Cleanup files
-/usr/bin/sudo rm -f /var/tmp/.hash
-/usr/bin/sudo rm -f ${sharedKxHome}/.config/.vmCredentialsFile
+       # Loop through encrypted credentials file and load into GoPass
+       while IFS='' read -r credential || [[ -n "$credential" ]]; do
+       name=$(echo ${credential} | cut -f 1 -d':')
+       secret=$(echo ${credential} | cut -f 2 -d':')
+       if [[ "${secret}" != '""' ]] && [[ "${secret}" != "" ]]; then
+              decryptedSecret=$(echo "${secret}" | openssl enc -aes-256-cbc -pbkdf2 -salt -A -a -pass pass:${hash} -d)
+              cleanedSecret=$(cleanOutput "${decryptedSecret}")
+       else
+              decryptedSecret=""
+       fi
+       pushPassword "${name}" "${cleanedSecret}" "base-technical-credentials"
+       done < "${sharedKxHome}/.config/.vmCredentialsFile"
+
+       # Cleanup files
+       /usr/bin/sudo rm -f /var/tmp/.hash
+       /usr/bin/sudo rm -f ${sharedKxHome}/.config/.vmCredentialsFile
+
+fi
